Napisan F1 Help za sve funkcionalnosti osim Planiranje (zadužena kolegica Fiorencis) i Statistika (zadužen kolega Ilić). Implementiran F1 na sve glavne forme. Napisani komentari tj. sažetci svih ručno pisanih funkcija i dodatnih konstruktora. Ukomponiran rad s opremom. Popravljeni manji bugovi.

+++ oplan/RadSPostrojbama.cs
@@ -2,6 +2,9 @@
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data.Entity;
+using System.Data.Entity.Core;
+using System.Data.Entity.Core.Objects;
+using System.Data.Entity.Infrastructure;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -14,8 +17,9 @@ namespace oplan
         static private postrojba postrojba = null;
 
         /// <summary>
-        /// Prikazuje popis postrojbi u glavnom prozoru.
+        /// Pomoću LINQ upita dohvaća i prikazuje popis postrojbi u glavnom prozoru.
         /// </summary>
+        /// <param name="dgvPostrojbe">Naziv DataGridViewa u kojem se prikazuju podaci</param>
         static public void PrikaziPostrojbe(DataGridView dgvPostrojbe)
         {
             using (var db = new EntitiesSettings())
@@ -42,6 +46,7 @@ namespace oplan
         /// <summary>
         /// Otvara novu formu za dodavanje nove postrojbe te prikazuje ažurirane postrojbe po završetku.
         /// </summary>
+        /// <param name="dgvPostrojbe">Naziv DataGridViewa u kojem se prikazuju podaci</param>
         static public void DodajPostrojbu(DataGridView dgvPostrojbe)
         {
             frmDodajPostrojbu formaPostrojba = new frmDodajPostrojbu();
@@ -50,8 +55,10 @@ namespace oplan
         }
 
         /// <summary>
-        /// Briše označenu postrojbu ako se ona ne kreće po niti jednoj ruti te prikazuje ažurirani popis postrojbi.
+        /// Briše označenu postrojbu ako se ona ne nalazi na ni jednom planu i nije joj dodjeljena oprema.
         /// </summary>
+        /// <param name="dgvPostrojbe">Naziv DataGridViewa u kojem se prikazuju podaci</param>
+        /// <param name="currentRow">Redak u kojem je postrojba koja se želi obrisati</param>
         static public void IzbrisiPostrojbu(DataGridView dgvPostrojbe, DataGridViewRow currentRow)
         {
             if (currentRow != null)
@@ -67,12 +74,20 @@ namespace oplan
                             {
                                 if (postrojba.tocka.Count == 0)
                                 {
-                                    db.postrojba.Remove(postrojba);
-                                    db.SaveChanges();
+                                    if (postrojba.oprema.Count == 0)
+                                    {
+                                        db.postrojba.Remove(postrojba);
+                                        MessageBox.Show("Uspješno ste obrisali postrojbu.", "Informacija", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                                    }
+                                    else
+                                    {
+                                        //OPCIONALNO: pitati dal se hoće obrisati i sve dodjele opreme u arsenalu i implementirati
+                                        MessageBox.Show("Nije moguće izbrisati postrojbu kojoj je dodjeljena oprema!", "Pogreška", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                                    }
                                 }
                                 else
                                 {
-                                    MessageBox.Show("Nije moguće izbrisati postrojbu koja se kreće po postojećoj ruti!", "Pogreška", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                                    MessageBox.Show("Nije moguće izbrisati postrojbu koja je na postojećem planu!", "Pogreška", MessageBoxButtons.OK, MessageBoxIcon.Error);
                                 }
                             }
                         }
@@ -85,7 +100,10 @@ namespace oplan
         /// <summary>
         /// Provjerava je li postrojba bila mijenjana tijekom izmjene te postoji li takva u bazi podataka.
         /// </summary>
-        /// <returns>True ako se postrojba može mijenjati, false ako se ne može.</returns>
+        /// <param name="idVrsta">ID vrste postrojbe koja se provjerava</param>
+        /// <param name="idTip">ID tipa postrojbe koja se provjerava</param>
+        /// <param name="redakZaIzmjenu">Redak u kojem je postrojba koja se želi provjeriti</param>
+        /// <returns>True ako se postrojba može unjeti, false ako se ne može.</returns>
         static public bool ProvjeriPostrojbu(int idVrsta, int idTip, DataGridViewRow redakZaIzmjenu)
         {
             bool promjena = false;
@@ -120,6 +138,8 @@ namespace oplan
         /// <summary>
         /// Pomoću upita sa ključem vrste i tipa provjerava postoji li takva postrojba u bazi.
         /// </summary>
+        /// <param name="idVrsta">ID vrste kojeg postrojba ima</param>
+        /// <param name="idTip">ID tipa kojeg je postrojba</param>
         /// <returns>True ako postrojba ne postoji, false ako postrojba postoji.</returns>
         static public bool ProvjeriPostojanjePostrojbe(int idVrsta, int idTip)
         {
@@ -142,6 +162,8 @@ namespace oplan
         /// <summary>
         /// Prikazuje formu za izmjenu postrojbe te prikazuje ažurirane postrojbe po završetku.
         /// </summary>
+        /// <param name="dgvPostrojbe">Naziv DataGridViewa u kojem se prikazuju podaci</param>
+        /// <param name="currentRow">Redak postrojbe koja se želi izmjeniti</param>
         static public void IzmijeniPostrojbu(DataGridView dgvPostrojbe, DataGridViewRow currentRow)
         {
             if (currentRow != null)
@@ -151,5 +173,45 @@ namespace oplan
                 PrikaziPostrojbe(dgvPostrojbe);
             }
         }
+
+        /// <summary>
+        /// Prikazuje formu sa izvještajem ako za tu postrojbu postoji dodjeljena oprema.
+        /// </summary>
+        /// <param name="id_postrojbe">ID postrojbe kojoj se želi vidjeti oprema</param>
+        /// <param name="naziv">Naziv postrojbe u tekstualnom obliku</param>
+        static public void PrikaziOpremu(int id_postrojbe, string naziv)
+        {
+            if (Izvjestaji.ProvjeriOpremu(id_postrojbe))
+            {
+                frmNaoruzanje popisOpreme = new frmNaoruzanje(id_postrojbe, naziv);
+                popisOpreme.ShowDialog();
+            }
+            else
+            {
+                MessageBox.Show("Ovoj postrojbi nije dodjeljena niti jedna oprema.", "Informacija", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
+        }
+
+        //Funkcija koja briše sve dodjele opreme određenoj postrojbi i trenutno nije aktivna.
+        /*public static void IzbrisiDodjele(int id_postrojba)
+        {
+            using (var db = new EntitiesSettings())
+            {
+                var upit = (from p in db.postrojba
+                            from o in p.oprema
+                            join s in db.oprema on o.id_oprema equals s.id_oprema
+                            where p.id_postrojba == id_postrojba
+                            select new
+                            {
+                                IDP = p.id_postrojba,
+                                IDO = s.id_oprema,
+                            }).ToList();
+
+                foreach (var item in upit)
+                {
+                    RadSArsenalom.ObrisiArsenal(item.IDP, item.IDO);
+                }
+            }
+        }*/
     }
 }
